+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### SNAPSHOT
 * Added ability to verify a download with a hash checksum
+* Added ability to delete downloads with their local files
 * Added ability to edit download properties
   * Changing location will cause the file to be moved, in a safe way and with a progress bar
 * Reworked download location selection in Add Download dialog